feat-22 refactor files
