+++ src/contexts/actionContext.tsx
@@ -1,0 +1,3 @@
+export enum ACTIONS {
+  SET_CATEGORIES = 'set_categories',
+}
